Rename and consolidate dashboard files
- Rename `fasthtml-dashboard.py` to `dashboard.py`
- Remove `streamlit-dashboard.py`
- Update Makefile to use new dashboard script names
- Update dashboard README to reflect current dashboard implementation
- Simplify dashboard utility functions and chart rendering
