+++ dashboard/README.md
@@ -1,6 +1,3 @@
 # Dashboard
 
-Some simple local dashboard options live here.
-
-- [`streamlit-dashboard.py`](streamlit-dashboard.py): [Streamlit](https://streamlit.io/) dashboard for the project.
-- [`fasthtml-dashboard.py`](fasthtml-dashboard.py): [FastHTML](https://fastht.ml/) dashboard for the project (WIP).
+- [`dashboard.py`](dashboard.py): [FastHTML](https://fastht.ml/) dashboard for the project (WIP).
